Refactoring - Made Ch 10 Silver Challenge its own module
This may not be a good design from an MVC perspective (we'll see in Ch 11 how the book designs DOM updates with modules, and we may refactor this again then).  But I separated it out into its own module anyway as a slight improvement, and so it can be toggled on and off as a feature.